+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ===============
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/9)
 * Optimialization using memo, useMemo and useCallback
-* Create first api call for carbon threshold
+* Create and implement api call for carbon thresholds
+* Create and implement api call for all companies
 
 0.18 (2024-05-21)
 ===============